Separate OpenSSL-dependent and -independent Parts of C++ Crypto Code
`common/cpp/crypto/`
- crypto_utils_encrypt.cpp:
  split OpenSSL-independent part from crypto_utils.cpp
- crypto_shared.h: Add CRYPTOLIB_OPENSSL as default
- *.h:   Remove use of OpenSSL-specific constants in public header files
- *.cpp: Remove unused header files
- Add Doxygen documentation
