+++ common/cpp/crypto/crypto_utils.cpp
@@ -13,42 +13,28 @@
  * limitations under the License.
  */
 
-#include <string.h>
-#include <assert.h>
+/**
+ * @file
+ * Avalon Crypto Utilities:
+ * hashing, base 64 conversion, random number generation.
+ * Implemented using OpenSSL.
+ */
+
 #include <openssl/err.h>
 #include <openssl/rand.h>
 #include <openssl/sha.h>
-#include <openssl/x509v3.h>
-#include <openssl/pem.h>
-#include <algorithm>
-#include <memory>
-#include <vector>
 
-#include "base64.h"  // Simple base64 enc/dec routines
 #include "crypto_shared.h"
 #include "crypto_utils.h"
 #include "error.h"
 #include "tcf_error.h"
-#include "hex_string.h"
-#include "c11_support.h"
-#include "crypto.h"
-#include "utils.h"
-
-/*** Conditional compile untrusted/trusted ***/
-#if _UNTRUSTED_
-#include <openssl/crypto.h>
-#include <stdio.h>
-
-#else
-#include "tSgxSSL_api.h"
-#endif
-/*** END Conditional compile untrusted/trusted ***/
 
 namespace pcrypto = tcf::crypto;
 namespace constants = tcf::crypto::constants;
 
 // Error handling
 namespace Error = tcf::error;
+
 
 /**
  * Generate a cryptographically strong random bit string.
@@ -91,62 +77,7 @@ ByteArray pcrypto::RandomBitString(size_t length) {
  * @returns byte array containing binary hash of data
  */
 ByteArray pcrypto::ComputeMessageHash(const ByteArray& message) {
-    ByteArray hash(SHA256_DIGEST_LENGTH);
+    ByteArray hash(constants::DIGEST_LENGTH);
     SHA256((const unsigned char*)message.data(), message.size(), hash.data());
     return hash;
 }  // pcrypto::ComputeMessageHash
-
-
-/**
- * Create symmetric encryption key and return hex-encoded key string.
- * Uses AES-GCM 256, which also includes authentication.
- * Key generated is 256 bits represented as a 64 hex digit printable string.
- * Key for use with EncryptData() and DecryptData().
- *
- * @returns AES-GCM-256 key as a hex-encoded string
- */
-std::string pcrypto::CreateHexEncodedEncryptionKey() {
-    ByteArray enc_key = tcf::crypto::skenc::GenerateKey();
-    return ByteArrayToHexEncodedString(enc_key);
-}   // pcrypto::CreateHexEncodedEncryptionKey
-
-
-/**
- * Encrypt the message using given encryption key and return cipher.
- * Uses AES-GCM 256, which also includes authentication.
- * Use symmetric encryption key generated by CreateHexEncodedEncryptionKey().
- * Implemented using pcrypto::skenc::EncryptMessage().
- *
- * @param msg binary data to encrypt
- * @param key symmetric encryption key in a hex-encoded string.
- *            Generated by CreateHexEncodedEncryptionKey()
- * @returns base64 encoded string containing encrypted data
- */
-std::string pcrypto::EncryptData(std::string msg, std::string key) {
-    ByteArray msg_bytes = StrToByteArray(msg);
-    ByteArray key_bytes = tcf::HexStringToBinary(key);
-    ByteArray cipher = tcf::crypto::skenc::EncryptMessage(key_bytes,
-        msg_bytes);
-    return ByteArrayToBase64EncodedString(cipher);
-}  // pcrypto::EncryptData
-
-
-/**
- * Decrypt cipher using given encryption key and return message.
- * Uses AES-GCM 256, which also includes authentication.
- * Use symmetric encryption key generated by CreateHexEncodedEncryptionKey().
- * Implemented using pcrypto::skenc::DecryptMessage().
- *
- * @param cipher encrypted binary data encoded as a base64 string.
- *               Generated by EncryptMessage()
- * @param key    symmetric encryption key in a hex-encoded string.
- *               Generated by CreateHexEncodedEncryptionKey()
- * @returns base64 encoded string containing encrypted data
- */
-std::string pcrypto::DecryptData(std::string cipher, std::string key) {
-    ByteArray ciphers_bytes = Base64EncodedStringToByteArray(cipher);
-    ByteArray key_bytes = tcf::HexStringToBinary(key);
-    ByteArray msg = tcf::crypto::skenc::DecryptMessage(key_bytes,
-        ciphers_bytes);
-    return ByteArrayToStr(msg);
-}  // pcrypto::DecryptData
