+++ common/cpp/crypto/pkenc.h
@@ -13,13 +13,12 @@
  * limitations under the License.
  */
 
-/*
- * Private header file for use by pkenc_*_key.cpp.
- * Avalon RSA definitions common for public/private keypair functions.
+/**
+ * @file
+ * Avalon RSA constant definitions.
  */
 
 #pragma once
-#include <openssl/rsa.h>
 
 namespace tcf {
 namespace crypto {
@@ -31,8 +30,6 @@ namespace crypto {
         const int RSA_KEY_SIZE = 2048;
         const int RSA_PADDING_SIZE = 41;
 
-        // OAEP or better should always be used for RSA encryption.
-        const int RSA_PADDING_SCHEME = RSA_PKCS1_OAEP_PADDING;
         constexpr int RSA_PLAINTEXT_LEN =
             ((RSA_KEY_SIZE - RSA_PADDING_SIZE) >> 3);
     }  // namespace constants
