+++ common/cpp/crypto/pkenc_public_key.cpp
@@ -12,35 +12,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "pkenc_public_key.h"
+
+/**
+ * @file
+ * Avalon RSA public key generation, serialization, and encryption functions.
+ */
+
 #include <openssl/err.h>
 #include <openssl/pem.h>
-#include <openssl/rand.h>
-#include <openssl/sha.h>
-#include <algorithm>
-#include <memory>
-#include <vector>
-#include "base64.h"  //simple base64 enc/dec routines
+#include <memory>    // std::unique_ptr
+
 #include "crypto_shared.h"
 #include "error.h"
 #include "hex_string.h"
 #include "pkenc.h"
 #include "pkenc_private_key.h"
 #include "pkenc_public_key.h"
-
-/***Conditional compile untrusted/trusted***/
-#if _UNTRUSTED_
-#include <openssl/crypto.h>
-#include <stdio.h>
-#else
-#include "tSgxSSL_api.h"
-#endif
-/***END Conditional compile untrusted/trusted***/
-
-/**
- * @file
- * Avalon RSA public key generation, serialization, and encryption functions.
- */
 
 namespace pcrypto = tcf::crypto;
 namespace constants = tcf::crypto::constants;
@@ -231,6 +218,7 @@ std::string pcrypto::pkenc::PublicKey::Serialize() const {
 
 /**
  * Encrypt message with RSA public key and return ciphertext.
+ * Uses PKCS1 OAEP padding.
  * Throws RuntimeError.
  *
  * @param message ByteArray containing raw binary plaintext
