+++ common/cpp/crypto/skenc.cpp
@@ -18,34 +18,16 @@
  * Uses AES-GCM 256, which also includes authentication.
  */
 
-#include "skenc.h"
-#include <openssl/err.h>
 #include <openssl/pem.h>
-#include <openssl/rand.h>
-#include <openssl/sha.h>
-#include <algorithm>
-#include <memory>
-#include <vector>
-#include "base64.h"  // simple base64 enc/dec routines
+#include <memory>    // std::unique_ptr
+
 #include "crypto_shared.h"
-#include "crypto_utils.h"
+#include "crypto_utils.h" // RandomBitString()
 #include "error.h"
 #include "hex_string.h"
 #include "skenc.h"
 
-/***Conditional compile untrusted/trusted***/
-
-#if _UNTRUSTED_
-
-#include <openssl/crypto.h>
-#include <stdio.h>
-#else
-#include "tSgxSSL_api.h"
-#endif
-
 namespace pcrypto = tcf::crypto;
-
-/***END Conditional compile untrusted/trusted***/
 
 // Typedefs for memory management
 // Specify type and destroy function type for unique_ptrs
@@ -70,7 +52,8 @@ ByteArray pcrypto::skenc::GenerateKey() {
 
 
 /**
- * Generate symmetric authenticated encryption Initialization Vector (IV).
+ * Generate a 96-bit symmetric authenticated encryption Initialization Vector
+ * (IV). An IV is sometimes called a nonce.
  * Throws RuntimeError.
  *
  * @param IVstring string to convert to a IV.
@@ -82,7 +65,7 @@ ByteArray pcrypto::skenc::GenerateIV(const std::string& IVstring) {
     if (IVstring.compare("") == 0)
         return pcrypto::RandomBitString(constants::IV_LEN);
     // else use IVstring
-    ByteArray hash(SHA256_DIGEST_LENGTH);
+    ByteArray hash(pcrypto::constants::DIGEST_LENGTH);
     SHA256((const unsigned char*)IVstring.data(), IVstring.size(), hash.data());
     hash.resize(constants::IV_LEN);
     return hash;
@@ -90,14 +73,19 @@ ByteArray pcrypto::skenc::GenerateIV(const std::string& IVstring) {
 
 
 /**
- * Encrypt a message using authenticated encryption with random IV
- * prepended to ciphertext.
+ * Encrypt a message using AES-GCM authenticated encryption.
+ *
+ * Prepends a 12 byte (96 bit) IV (sometimes called a nonce) and
+ * appends a 16 byte (128 bit) authentication tag (sometimes called a MAC)
+ * to the output cipher text:
+ *     message = IV + ciphertext + authentication tag
+ * The IV and nonce are not encrypted.
  * Throws RuntimeError, ValueError.
  *
  * @param message binary data to encrypt
  * @param key Secret AES-256 encryption key.
  *            Generated by GenerateKey()
- * @param iv  Initialization Vector (IV). Generated by GenerateIV()
+ * @param iv  96-bit initialization Vector (IV). Generated by GenerateIV()
  * @returns Byte array containing encrypted data
  */
 ByteArray pcrypto::skenc::EncryptMessage(
@@ -186,7 +174,8 @@ ByteArray pcrypto::skenc::EncryptMessage(
 
 
 /**
- * Encrypt a message using authenticated encryption.
+ * Encrypt a message using AES-GCM authenticated encryption.
+ * An IV is generated automatically.
  * Throws RuntimeError, ValueError.
  *
  * @param message binary data to encrypt
@@ -216,8 +205,15 @@ ByteArray pcrypto::skenc::EncryptMessage(const ByteArray& key,
 
 
 /**
- * Decrypt message.data() using authenticated decryption.
- * Throws RuntimeError, ValueError, CryptoError.
+ * Decrypt message.data() using AES-GCM authenticated decryption.
+ *
+ * Expects an appended 12 byte (96 bit) IV (sometimes called a nonce) and
+ * a prepended 16 byte (128 bit) authentication tag (sometimes called a MAC)
+ * prepended and appended, respectively, the input cipher text:
+ *     message = IV + ciphertext + authentication tag
+ * The IV and nonce are not encrypted and were were added by EncryptMessage().
+ * Throws RuntimeError, ValueError,
+ * CryptoError (message authentication failure).
  *
  * @param message binary data to decrypt. Generated by EncryptMessage()
  * @param key Secret AES-256 encryption key.
@@ -310,12 +306,15 @@ ByteArray pcrypto::skenc::DecryptMessage(
 
 
 /**
- * Decrypt message.data() using authenticated encryption.
- * Expects IV prepended to message ciphertext and authentication tag
- * appended to cipher text:
- * message = IV + ciphertext + authentication tag
- * IV was prepended by EncryptMessage().
- * Throws RuntimeError, ValueError.
+ * Decrypt message.data() using AES-GCM authenticated encryption.
+ *
+ * Expects an appended 12 byte (96 bit) IV (sometimes called a nonce) and
+ * a prepended 16 byte (128 bit) authentication tag (sometimes called a MAC)
+ * prepended and appended, respectively, the input cipher text:
+ *     message = IV + ciphertext + authentication tag
+ * The IV and nonce are not encrypted and were were added by EncryptMessage().
+ * Throws RuntimeError, ValueError,
+ * CryptoError (message authentication failure).
  *
  * @param message binary data to decrypt. Generated by EncryptMessage()
  * @param key Secret AES-256 encryption key.
@@ -341,8 +340,7 @@ ByteArray pcrypto::skenc::DecryptMessage(const ByteArray& key,
     if (message.size() < constants::IV_LEN + constants::TAG_LEN) {
         std::string msg(
             "Crypto Error (DecryptMessage): AES-GCM message smaller "
-            "than minimum length (IV length + TAG "
-            "length)");
+            "than minimum length (IV length + TAG length)");
         throw Error::ValueError(msg);
     }
